NL formatter class
